backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.3
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
email-validator==2.2.0
PyYAML==6.0.2
python-dotenv==1.1.0
httpx==0.28.1
aiofiles==24.1.0
stripe==12.0.0
pandas==2.2.3
requests==2.32.3
investpy==1.0.8
setuptools<82
eth-account==0.13.5
web3==7.8.0
prometheus-client==0.21.1